Using context to manage global filter component
